Add an option 'sterile' to the model:
Cooperators never reproduce when this option is turned on.

If a neighborhood contains only cooperators and thus no individual that
could potentially reproduce, we quit with an error.

The cost option is still used, since it is somehow a cost/benefit ratio.
Its value thus still contributes to the competition between non
cooperators.

So setting cost=1 is **not the same** than this option.